stream: Remove "simple" stream API
The simple stream API is a way to create a stream from just a mainloop.
Context and Core are handled automatically internally.

Accomodating this API severely complicates the code for our rust bindings,
making them messy and much bigger.

Since the simple API only saves a couple lines for creating a context and core,
its benefit is not very much, so this commit removes the simple API and simplifies the stream code.