stream: Accept `StreamRef` and `D` in all callbacks
It doesn't seem like there's any reason not to provide these to all the
callbacks. In some cases they're useful (and require ugly hacks to work
around their absence), so it seems best to just be consistent.

This also changes the order of arguments to `param_changed` to have the
param id after `D`, again for consistency.