pipewire: Make PropertiesRef struct into a transparent wrapper
Before this commit, PropertiesRef was not a transparent wrapper like the other ref structs in the crate are,
so you couldn't use it in places were you just had a pointer to a nonowned pw_sys::pw_properties.

Deref/DerefMut impls are added to the Properties struct so that it can be converted automatically as needed.